requirements.txt
certifi
chardet
idna
requests
selenium
webdriver-manager
pexels-api
fastapi
uvicorn
python-multipart
packaging
python-dotenv